Revert "Create a function to avoid duplication of code"
This reverts commit 74fd84b484beb8b891ee028a3f5c49f7f8047f16.
