+++ script/src/syscalls/exec_v2.rs
@@ -1,6 +1,6 @@
-use crate::syscalls::utils::validate_offset_length;
 use crate::syscalls::{
-    Place, Source, EXEC, INDEX_OUT_OF_BOUND, SOURCE_ENTRY_MASK, SOURCE_GROUP_FLAG,
+    Place, Source, EXEC, INDEX_OUT_OF_BOUND, SLICE_OUT_OF_BOUND, SOURCE_ENTRY_MASK,
+    SOURCE_GROUP_FLAG,
 };
 use crate::types::{DataLocation, DataPieceId, ExecV2Args, Message, TxData, VmId};
 use ckb_traits::{CellDataProvider, ExtensionProvider, HeaderProvider};
@@ -74,15 +74,32 @@ where
         let offset = bounds >> 32;
         let length = bounds as u32 as u64;
 
-        // We are fetching the actual cell here for some in-place validation.
-        if !validate_offset_length(
-            machine,
-            self.snapshot2_context.clone(),
-            &data_piece_id,
-            offset,
-            length,
-        )? {
+        // We are fetching the actual cell here for some in-place validation
+        let mut sc = self
+            .snapshot2_context
+            .lock()
+            .map_err(|e| VMError::Unexpected(e.to_string()))?;
+        let (_, full_length) = match sc.load_data(&data_piece_id, 0, 0) {
+            Ok(val) => val,
+            Err(VMError::SnapshotDataLoadError) => {
+                // This comes from TxData results in an out of bound error, to
+                // mimic current behavior, we would return INDEX_OUT_OF_BOUND error.
+                machine.set_register(A0, Mac::REG::from_u8(INDEX_OUT_OF_BOUND));
+                return Ok(true);
+            }
+            Err(e) => return Err(e),
+        };
+        if offset >= full_length {
+            machine.set_register(A0, Mac::REG::from_u8(SLICE_OUT_OF_BOUND));
             return Ok(true);
+        }
+        if length > 0 {
+            // Both offset and length are <= u32::MAX, so offset.checked_add(length) will be always a Some.
+            let end = offset.checked_add(length).ok_or(VMError::MemOutOfBound)?;
+            if end > full_length {
+                machine.set_register(A0, Mac::REG::from_u8(SLICE_OUT_OF_BOUND));
+                return Ok(true);
+            }
         }
 
         let argc = machine.registers()[A4].to_u64();
